remove explicit boxing code
git-svn-id: https://svn.exedio.com/cope/trunk/persistence@8635 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ instrument/testsrc/com/exedio/cope/instrument/testmodel/Input.java
@@ -107,7 +107,7 @@ public class Input extends Item
 	 */
 	public static final Input findByUnique(final java.lang.String string,final int integer)
 	{
-		return Input.unique.searchUnique(Input.class,string,java.lang.Integer.valueOf(integer));
+		return Input.unique.searchUnique(Input.class,string,integer);
 	}/**
 
 	 **
