+++ runtime/testmodelsrc/com/exedio/cope/testmodel/ItemWithDoubleUnique.java
@@ -137,7 +137,7 @@ public class ItemWithDoubleUnique extends Item
 	 */
 	public static final ItemWithDoubleUnique findByDoubleUnique(final java.lang.String string,final int integer)
 	{
-		return ItemWithDoubleUnique.doubleUnique.searchUnique(ItemWithDoubleUnique.class,string,java.lang.Integer.valueOf(integer));
+		return ItemWithDoubleUnique.doubleUnique.searchUnique(ItemWithDoubleUnique.class,string,integer);
 	}/**
 
 	 **
